LibTLS: Move `DefaultRootCACertificates` to correct header file
The declaration of `DefaultRootCACertificates` was in `Certificate.h`
and its implementation in `TLSv12.cpp`. It has been moved over
to `TLSv12.h` for consistency.

This is in preparation of the next commits to split the changes.

(cherry picked from commit e42410a7a7e444966a3491382e077feff174a7a5)
