+++ Userland/Libraries/LibTLS/Certificate.h
@@ -300,24 +300,6 @@ private:
     Optional<bool> m_is_self_signed;
 };
 
-class DefaultRootCACertificates {
-public:
-    DefaultRootCACertificates();
-
-    Vector<Certificate> const& certificates() const { return m_ca_certificates; }
-
-    static ErrorOr<Vector<Certificate>> parse_pem_root_certificate_authorities(ByteBuffer&);
-    static ErrorOr<Vector<Certificate>> load_certificates(Span<ByteString> custom_cert_paths = {});
-
-    static DefaultRootCACertificates& the();
-
-    static void set_default_certificate_paths(Span<ByteString> paths);
-
-private:
-    Vector<Certificate> m_ca_certificates;
-};
-
 }
 
 using TLS::Certificate;
-using TLS::DefaultRootCACertificates;
